project_mind/requirements_txt.txt
streamlit==1.32.0
langchain==0.1.16
langchain-openai==0.1.1
langgraph==0.0.38
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
chromadb==0.4.24
tiktoken==0.6.0
openai==1.14.0
networkx==3.2.1
pandas==2.2.1
numpy==1.26.4
requests==2.31.0